Made some changes, in extended description
Incorporated mobs into one section
Environment damage is arena
Night/Day is arena
Potions is arena, under kits
Teams are arena - colors
Team size will be generated based on max players
Rewards are arena
Kits will be in mini game file, but should be able to be overridden in arena file


NOTE FOR MAX PLAYERS - this is in the arena file. If set to 0, there will be no limit. Display on signs will be with 3 more players than the current amount. Ex: 0/3, 19/22, 1838248/1838251, 1/4, 20/23, et cetera. Perhaps start with the minimum players. So 0/10, 6/10, 8/11, 19/22, etc.